tests/wpt/tests/custom-elements/behaves-like-button.tentative.html

<!DOCTYPE HTML>
<html>

<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>
  <script src="/resources/testdriver-actions.js"></script>
  <script src="/wai-aria/scripts/aria-utils.js"></script>
</head>

<body>
  <fancy-button popovertarget="myPopover" id="myFancyButton"></fancy-button>
  <div id="myPopover" popover>
    popover invoked by autonomous custom element with ElementInternals.type=button
  </div>

  <script>
    class FancyButton extends HTMLElement {
      constructor() {
        super();
        this.internals_ = this.attachInternals();
      }

      setType(type) {
        this.internals_.type = type;
      }
    }
    customElements.define('fancy-button', FancyButton);

    test(function () {
      const popover = document.getElementById("myPopover");

      let showCount = 0;
      popover.addEventListener('beforetoggle', (e) => {
        if (e.newState === "open") {
          ++showCount;
        }
      });

      const fancyButton = document.getElementById("myFancyButton");
      // Without type=button, popovertarget should not work.
      fancyButton.click();
      assert_equals(showCount, 0, "The popover should not be shown when the custom element doesn't have type=button.");

      // With type=button, popovertarget should work.
      fancyButton.setType("button");
      fancyButton.click();
      assert_equals(showCount, 1, "The popover should be shown once when the custom element has type=button.");
    }, "Autonomous custom element with ElementInternals.type=button supports the popovertarget attribute.");
  </script>
</body>

</html>
